refactor: Refactor planStatusStyles
